Add files via upload
We will calculate the student GPA and display the results using txt file. Student GPA will be calculated through a process which explains if the student is a freshman, junior, etc and how many courses plus the credit hours they are enrolled in and finally, what grade do they have in each course. Next we will get the GPA.